small fixes to configuration
