+++ CompuCell3D/player/Configuration/SettingUtils.py
@@ -3,6 +3,8 @@ from PyQt4.QtCore import *
 from os import environ,path
 import os
 
+from DefaultSettingsData import *
+
 def writeSettings (settingsObj,path):
     if settingsObj:
         settingsObj.saveAsXML(path)   
@@ -21,7 +23,57 @@ def loadSettings(filename):
         
     return None 
    
+def loadGlobalSettings():
+    
+    global_setting_dir = os.path.abspath(os.path.join(os.path.expanduser('~') , SETTINGS_FOLDER))
+    global_setting_path = os.path.abspath(os.path.join(global_setting_dir , SETTINGS_FILE_NAME)) # abspath normalizes path
+    # print 'LOOKING FOR global_setting_path=',global_setting_path
+    
+    #create global settings  directory inside user home directory
+    if not os.path.isdir(global_setting_dir):
+        try:
+            os.makedirs(global_setting_dir)
+    
+        except:
+            print 'Cenfiguration: COuld not make directory: ',global_setting_dir, ' to store global settings. Please make sure that you have appropriate write permissions'
+            import sys
+            sys.exit()
+    
+    globalSettings = loadSettings (global_setting_path)  
+    
+    if not globalSettings:
+        globalSettings, default_setting_path = loadDefaultSettings()
+        globalSettings.saveAsXML(global_setting_path)        
         
+        return globalSettings , global_setting_path     
+    return  globalSettings , global_setting_path            
+
+def loadDefaultSettings():
+       
+    default_setting_path = os.path.abspath(os.path.join(os.path.dirname(__file__) , SETTINGS_FILE_NAME)) # abspath normalizes path
+
+    defaultSettings = loadSettings (default_setting_path)  
+    
+    if not defaultSettings:
+        return None, None
+        
+    return defaultSettings , default_setting_path
+    
+    
+#this function checks for new settings in the default settings file    
+def synchronizeGlobalAndDefaultSettings(defaultSettings,globalSettings,globalSettingsPath):
+    defaultSettingsNameList = defaultSettings.getSettingNameList()
+    globalSettingsNameList = globalSettings.getSettingNameList()
+    
+    newSettingNames = set(defaultSettingsNameList) - set(globalSettingsNameList)
+    
+    for newSettingName in newSettingNames:
+        print 'newSettingName = ', newSettingName
+        newDefaultSetting = defaultSettings.getSetting(newSettingName)
+        
+        globalSettings.setSetting(newDefaultSetting.name , newDefaultSetting.value , newDefaultSetting.type)
+        writeSettings (globalSettings,globalSettingsPath) 
+    # sys.exit()        
         
 class Setting(object):
     storedType2XML = {
